twl92230: change pwrbtn_state to uint8_t
its value is always the level of an interrupt, 0 or 1

+++ hw/twl92230.c
@@ -62,7 +62,7 @@ typedef struct {
     } rtc;
     qemu_irq out[4];
     qemu_irq *in;
-    int pwrbtn_state;
+    uint8_t pwrbtn_state;
     qemu_irq pwrbtn;
 } MenelausState;
